+++ controllers/front/callbackopen.php
@@ -8,7 +8,6 @@
  */
 
 require_once _PS_MODULE_DIR_.'/altapay/lib/altapay/altapay-php-sdk/lib/AltapayCallbackHandler.class.php';
-require_once _PS_MODULE_DIR_.'/altapay/helpers.php';
 
 class AltapayCallbackopenModuleFrontController extends ModuleFrontController
 {