+++ CHANGELOG.md
@@ -5,6 +5,13 @@ PrestaShop 1.6.x and 1.7.x
 
 # Changelog
 
+## [3.3.1] - 2020-12-03
+### Added
+ - Support for payment captures full reservation after discount is applied from the backend
+
+ ### Fixed
+ - Fix shipping orderlines issue in full capture
+
 ## [3.3.0] - 2020-11-24
 ### Changed
  - Modify licence file with MIT License